[FLINK-4664] [gelly] Add translator to NullValue
This translator is appropriate for translating vertex and edge values to
NullValue when the values are not used by an algorithm. Also, added
tests and moved translators into a subpackage.

This closes #2536
